Issue #3954: MethodParamPad Test input split
